[libzstd] Speed up single segment zstd_fast by 5%
This PR is based on top of PR #1563.

The optimization is to process two input pointers per loop.
It is based on ideas from [igzip] level 1, and talking to @gbtucker.

| Platform                | Silesia     | Enwik8 |
|-------------------------|-------------|--------|
| OSX clang-10            | +5.6%       | +7.5%  |
| i9 5 GHz gcc-8          | +4.7%       | +6.8%  |
| i9 5 GHz clang-7        | +5.5%       | +7.9%  |
| Skylake 2.4 GHz gcc-4.8 | +4.8%       | +5.6%  |
| Skylake 2.4 GHz clang-7 | +6.0%       | +8.5%  |

We lose 0.4% of compression ratio on Silesia.
We gain 0.2% of compression ratio on enwik8.
I also tested on the GitHub and hg-commands datasets without a dictionary,
and we gain a small amount of compression ratio on each, as well as speed.

[igzip]: https://github.com/01org/isa-l/tree/master/igzip