[API][lib] The ZSTD_NO_MALLOC macro hides allocating functions
When the ZSTD_NO_MALLOC macro is defined functions that require
malloc are removed. Comments explain which functions are removed,
and what the alternatives are.

This is useful for freestanding mode, where we will remove sections
of the API that require malloc/free.

The unnecessary function definitions are also hidden behind the macro.